tensilelite host code : float bpe and make it compatible with rocisa

+++ projects/hipblaslt/tensilelite/include/Tensile/ContractionProblem.hpp
@@ -670,7 +670,7 @@ namespace TensileLite
             return m_betaType;
         }
 
-        size_t computeTypeElementSize() const
+        float computeTypeElementSize() const
         {
             return DataTypeInfo::Get(m_betaType).elementSize;
         }
